Update styling, add Routes, add comments
