+++ src/components/SymptomManager/SymptomManager.css
@@ -1,0 +1,92 @@
+* {
+  box-sizing: border-box;
+}
+
+.symptom-manager-wrapper {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+  width: 100vw;
+  height: 100vh;
+  background-color: var(--secondary);
+
+  /* hide scrollbar but allow scrolling */
+  /* for Internet Explorer, Edge */
+  /* -ms-overflow-style: none; */
+  /* for Firefox */
+  /* scrollbar-width: none; */
+  overflow-y: scroll;
+}
+
+/* hide scroll bar for Chrome, Safari, and Opera */
+.symptom-manager-wrapper::-webkit-scrollbar {
+  /* display: none; */
+}
+
+.helper-buttons-wrapper {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-evenly;
+  align-items: center;
+  width: 80%;
+  max-width: 630px;
+  margin: 110px 0 10px 0;
+  /* outline: 2px solid red; */
+}
+
+.delete-all-btn,
+.generate-cards-helper-btn,
+.sort-by-date-btn {
+  /* margin: 20px auto; */
+  padding: 5px 10px;
+  font-size: 1rem;
+  font-weight: 600;
+  color: #fff;
+  border-radius: 6px;
+  border: none;
+  cursor: pointer;
+}
+
+.delete-all-btn {
+  background-color: #af0303;
+}
+
+.generate-cards-helper-btn {
+  background-color: #089208;
+}
+
+.sort-by-date-btn {
+  width: 220px;
+  background-color: #636363;
+}
+
+.sort-order-message,
+.empty-message {
+  width: 100%;
+  height: fit-content;
+  margin: 15px auto 0;
+  font-size: 1.5rem;
+  font-weight: 400;
+  text-align: center;
+}
+
+.symptom-card-component-list-wrapper {
+  display: flex;
+  /* //* column-reverse shows newest symptom cards first */
+  flex-direction: column-reverse;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: fit-content;
+}
+
+/* Utility Class for for sorting cards by date descending / ascending */
+
+.descending-date {
+  flex-direction: column-reverse;
+}
+.ascending-date {
+  flex-direction: column;
+}
