+++ src/components/SymptomTracker/SymptomTracker.js
@@ -1,6 +1,9 @@
 //! This component is no longer in use
 //! This component is no longer in use
 //! This component is no longer in use
+//! This component is no longer in use
+//! This component is no longer in use
+//! This component is no longer in use
 
 import React from "react";
 import { useState } from "react";
