+++ src/index.js
@@ -6,14 +6,17 @@ import reportWebVitals from "./reportWebVitals";
 import { store, persistor } from "./redux/store";
 import { Provider } from "react-redux";
 import { PersistGate } from "redux-persist/integration/react";
+import { BrowserRouter, HashRouter } from "react-router-dom";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
     <Provider store={store}>
-      <PersistGate loading={"Loading"} persistor={persistor}>
-        <App />
-      </PersistGate>
+      <BrowserRouter>
+        <PersistGate loading={"Loading"} persistor={persistor}>
+          <App />
+        </PersistGate>
+      </BrowserRouter>
     </Provider>
   </React.StrictMode>
 );
